refactor: Category.java 파일 이동
