+++ src/main/java/com/team05/linkup/domain/enums/Category.java
@@ -1,12 +1,12 @@
-package com.team05.linkup.domain.community.domain;
+package com.team05.linkup.domain.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CommunityCategory {
-    QUESTION, // ("질문/답변"), // Keep korean names if needed for display elsewhere
+public enum Category {
+    QUESTION, // ("질문/답변"),
     INFO,     // ("정보공유"),
     REVIEW,   // ("후기"),
     FREE,     // ("자유게시판"),
